adsb-network-test/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
redis==5.0.1
psycopg2-binary==2.9.9
jinja2==3.1.2
websockets==12.0
